python/fastapi-tortoise-psql/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["manupanand <manupanandindianinvestor@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.6"
python-dotenv = "^1.0.1"
tortoise-orm = {extras = ["asyncpg"], version = "^0.23.0"}
asyncpg = "^0.30.0"
uvicorn = "^0.34.0"
pymongo = {extras = ["srv"], version = "^4.10.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
